requirements.txt
cmake
pyarrow==11.0.0
pandas
mlflow==2.2.2
notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyyaml
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
setuptools
-e .